Increase maxBuffer size for `cordova run ...` execution.
In order to accommodate a larger amount of output from Xcode, this
increases the default Node.js maxBuffer setting (200KB) to 5MB, which
should be more than sufficient for most build output.

In an ideal fix, this would switch to using Node.js' "spawn", which is
better suited for cases where the output of a command might be
extensive or long-running.  As a matter of urgency, I'm just going with
this.